requirements.txt
pytest
requests
python-dotenv-1.2.2
allure-pytest